python/requirements.txt
gymnasium==0.29.1
stable-baselines3[extra]==2.3.2
numpy==1.26.4
torch==2.2.2
